crawler/requirements.txt
selenium==4.21.0
webdriver-manager==4.0.1
beautifulsoup4==4.12.3
requests==2.31.0
python-dotenv
